Address PR review: use long for keyspace counts, poll for TTL expiry
- Switch keyspace key/expire counts from int to long end-to-end
  (scan counters, StorageSession.KeyspaceStats, IDatabaseManager /
  DatabaseManagerBase / Single+MultiDatabaseManager / StoreWrapper) so
  INFO KEYSPACE can safely report keyspaces larger than int.MaxValue.
- Replace the fixed Thread.Sleep(250) in InfoKeyspaceExpiredKeysNotCountedTest
  with a bounded poll on KeyExists to avoid CI flakiness.

+++ libs/server/Databases/DatabaseManagerBase.cs
@@ -598,7 +598,7 @@ namespace Garnet.server
         }
 
         /// <inheritdoc/>
-        public abstract (int keyCount, int expireCount) GetKeyspaceStats(int dbId);
+        public abstract (long keyCount, long expireCount) GetKeyspaceStats(int dbId);
 
         /// <summary>
         /// Count live keys and live keys with an expiration set for the given database via a full store scan.
@@ -607,7 +607,7 @@ namespace Garnet.server
         /// </summary>
         /// <param name="db">Database to scan.</param>
         /// <returns>A tuple of (live keys, live keys with an expiration set).</returns>
-        protected (int keyCount, int expireCount) GetDatabaseKeyspaceStats(GarnetDatabase db)
+        protected (long keyCount, long expireCount) GetDatabaseKeyspaceStats(GarnetDatabase db)
         {
             lock (db.KeyspaceScanLock)
             {
