+++ libs/server/Databases/MultiDatabaseManager.cs
@@ -1091,7 +1091,7 @@ namespace Garnet.server
         public override (long numExpiredKeysFound, long totalRecordsScanned) ExpiredKeyDeletionScan(int dbId)
             => StoreExpiredKeyDeletionScan(GetDbById(dbId));
 
-        public override (int keyCount, int expireCount) GetKeyspaceStats(int dbId)
+        public override (long keyCount, long expireCount) GetKeyspaceStats(int dbId)
             => GetDatabaseKeyspaceStats(GetDbById(dbId));
 
         private GarnetDatabase GetDbById(int dbId)
